dev: update deps, and move js to packages/admin
