+++ package.json
@@ -20,7 +20,7 @@
 		"lint:css-fix": "stylelint '**/*.scss' --fix",
 		"lint:js": "wp-scripts lint-js --ext=js,ts,tsx",
 		"lint:js:report": "npm run lint:js -- --output-file eslint_report.json --ext=js,ts,tsx --format json",
-		"lint:js-fix": "eslint src/assets/js --ext=js,jsx,ts,tsx --fix",
+		"lint:js-fix": "eslint packages/admin --ext=js,jsx,ts,tsx --fix",
 		"lint:php": "composer run-script check-cs",
 		"lint:php-fix": "composer run-script fix-cs",
 		"lint:phpstan": "composer run-script phpstan",
@@ -49,56 +49,56 @@
 		]
 	},
 	"dependencies": {
-		"@wordpress/i18n": "^4.34.0",
+		"@wordpress/i18n": "^4.40.0",
 		"uuid": "^9.0.0"
 	},
 	"devDependencies": {
-		"@babel/cli": "^7.21.5",
-		"@babel/core": "^7.21.8",
-		"@babel/plugin-syntax-jsx": "^7.21.4",
-		"@svgr/webpack": "^8.0.1",
+		"@babel/cli": "^7.22.10",
+		"@babel/core": "^7.22.10",
+		"@babel/plugin-syntax-jsx": "^7.22.5",
+		"@svgr/webpack": "^8.1.0",
 		"@types/classnames": "^2.3.1",
-		"@types/eslint": "^8.40.0",
-		"@types/lodash": "^4.14.195",
+		"@types/eslint": "^8.44.2",
+		"@types/lodash": "^4.14.197",
 		"@types/wordpress__components": "^23.0.1",
 		"@types/wordpress__notices": "^3.27.0",
-		"@typescript-eslint/eslint-plugin": "^5.59.7",
-		"@typescript-eslint/parser": "^5.59.7",
-		"@wordpress/babel-plugin-import-jsx-pragma": "^4.17.0",
-		"@wordpress/babel-preset-default": "^7.18.0",
-		"@wordpress/core-data": "^6.11.0",
-		"@wordpress/env": "^8.0.0",
-		"@wordpress/eslint-plugin": "^14.7.0",
-		"@wordpress/hooks": "^3.34.0",
-		"@wordpress/postcss-plugins-preset": "^4.18.0",
-		"@wordpress/prettier-config": "^2.17.0",
-		"@wordpress/readable-js-assets-webpack-plugin": "^2.17.0",
-		"@wordpress/scripts": "^26.5.0",
-		"@wordpress/stylelint-config": "^21.17.0",
+		"@typescript-eslint/eslint-plugin": "^5.62.0",
+		"@typescript-eslint/parser": "^5.62.0",
+		"@wordpress/babel-plugin-import-jsx-pragma": "^4.23.0",
+		"@wordpress/babel-preset-default": "^7.24.0",
+		"@wordpress/core-data": "^6.17.0",
+		"@wordpress/env": "^8.6.0",
+		"@wordpress/eslint-plugin": "^15.0.0",
+		"@wordpress/hooks": "^3.40.0",
+		"@wordpress/postcss-plugins-preset": "^4.24.0",
+		"@wordpress/prettier-config": "^2.23.0",
+		"@wordpress/readable-js-assets-webpack-plugin": "^2.23.0",
+		"@wordpress/scripts": "^26.11.0",
+		"@wordpress/stylelint-config": "^21.23.0",
 		"ajx": "^0.1.0",
-		"babel-loader": "^9.1.2",
+		"babel-loader": "^9.1.3",
 		"babel-plugin-inline-json-import": "^0.3.2",
 		"cross-env": "^7.0.3",
 		"cssnano": "^6.0.1",
 		"deep-freeze": "^0.0.1",
-		"eslint-import-resolver-node": "^0.3.7",
-		"eslint-import-resolver-webpack": "^0.13.2",
+		"eslint-import-resolver-node": "^0.3.9",
+		"eslint-import-resolver-webpack": "^0.13.6",
 		"eslint-plugin-eslint-comments": "^3.2.0",
-		"eslint-plugin-import": "^2.27.5",
+		"eslint-plugin-import": "^2.28.0",
 		"eslint-plugin-you-dont-need-lodash-underscore": "^6.12.0",
 		"glob": "^8.1.0",
-		"postcss": "^8.4.23",
-		"postcss-loader": "^7.3.1",
+		"postcss": "^8.4.28",
+		"postcss-loader": "^7.3.3",
 		"progress": "^2.0.3",
-		"rimraf": "^4.4.1",
-		"sass": "^1.62.1",
-		"sass-loader": "^13.3.0",
+		"rimraf": "^5.0.1",
+		"sass": "^1.66.0",
+		"sass-loader": "^13.3.2",
 		"terser-webpack-plugin": "^5.3.9",
-		"typescript": "^4.9.5",
+		"typescript": "^5.1.6",
 		"url-loader": "^4.1.1",
-		"webpack": "^5.84.1",
-		"webpack-bundle-analyzer": "^4.8.0",
-		"webpack-cli": "^5.1.1",
+		"webpack": "^5.88.2",
+		"webpack-bundle-analyzer": "^4.9.0",
+		"webpack-cli": "^5.1.4",
 		"webpack-rtl-plugin": "^2.0.0"
 	},
 	"overrides": {
