+++ packages/admin/settings/PluginSetings/JwtSecretControl.tsx
