+++ tsconfig.base.json
@@ -6,16 +6,21 @@
 		"rootDir": ".",
 		"sourceMap": true,
 		"strictNullChecks": true,
+		"incremental": true,
+		"noErrorTruncation": true,
 
 		"allowJs": true,
-		"allowSyntheticDefaultImports": true,
 		"checkJs": true,
+		"allowSyntheticDefaultImports": true,
+		"jsx": "preserve",
+		"target": "esnext",
+		"module": "esnext",
+		"lib": [ "dom", "esnext" ],
+		"declaration": true,
+		"declarationMap": true,
 		"composite": true,
 		"emitDeclarationOnly": true,
-		"jsx": "preserve",
-		"lib": [ "dom", "esnext" ],
-		"module": "esnext",
-		"target": "esnext",
+		"isolatedModules": true,
 
 		/* Strict Type-Checking Options */
 		"strict": true,
@@ -23,16 +28,16 @@
 		/* Additional Checks */
 		"noUnusedLocals": true,
 		"noUnusedParameters": true,
+		"noImplicitThis": true,
 		"noImplicitReturns": true,
 		"noFallthroughCasesInSwitch": true,
-		"importsNotUsedAsValues": "error",
 
 		/* Module Resolution Options */
 		"moduleResolution": "node",
 
-		/* This needs to be false so our types are possible to consume without setting this */
-		"esModuleInterop": false, // true in woo-blocks
+		"esModuleInterop": true,
 		"resolveJsonModule": true,
+
 		"typeRoots": [ "./node_modules/@types" ],
 		"types": [],
 	},
